requirements.txt
asgiref==3.8.1
Django==5.1.6
sqlparse==0.5.3
django-allauth == 65.3.0
django-allauth[socialaccount]
requests==2.32.3
channels==4.0.0
daphne==4.0.0
python-magic==0.4.27
python-magic-bin==0.4.14